Rename Windows-specific files
